Created Utils Folder and Minor Updates
